// src/models/Teachers/ITeacherWithInfoRow.ts
export interface ITeacherWithInfoRow {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    dateOfBirth: string;
    gender: string;
    signedUpAt: string;
    numberOfSubjects: number;
    numberOfClasses: number;
    position: string;
    description: string;
    experience: string;
}